DataAggregateInfra/app2/requirements.txt
requests
psycopg2-binary
pandas
websocket-client
pybit